+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix anthropic "invalid_request_error" that may happen when doing server web-search + thinking, causing errors in chat.
 - Allow fetching models from providers with no api keys required. #313
+- Fix usage tokens for anthropic when doing server web-searches. #307
 
 ## 0.101.1